feature: Added post delete funtionality and converted route logic to middleware
