+++ lib/controllers/users.js
@@ -1,11 +1,16 @@
 const jwt = require("jsonwebtoken");
 const dotenv = require('dotenv');
+const {
+  authenticateToken,
+  hashpassword,
+  validatepassword,
+} = require("../utils/auth_utils");
 
 dotenv.config();
 
 const secret = process.env.TOKEN_SECRET;
 
-const author = (authenticateToken, get_db) => {
+const author = (get_db) => {
   return function (req, res) {
     const db = get_db();
     const { token } = req.body;
@@ -27,7 +32,7 @@ const author = (authenticateToken, get_db) => {
   };
 };
 
-const register = (hashpassword, get_db) => {
+const register = (get_db) => {
   return async function (req, res) {
     if (req.body) {
       const db = get_db();
@@ -62,7 +67,7 @@ const register = (hashpassword, get_db) => {
   };
 };
 
-const login = (validatepassword, get_db) => {
+const login = (get_db) => {
   return async function (req, res) {
     const db = get_db();
